https

+++ app/Jobs/SuckMovieJob.php
@@ -36,6 +36,11 @@ class SuckMovieJob implements ShouldQueue
      */
     public function handle()
     {
+        $is_recent = Movie::where('id', $this->id)
+        ->where('updated_at', '>', Carbon::now()->subHours(30)->toDateTimeString())
+        ->first();
+        if($is_recent) return;
+
         if($this->isWithRecommendation){
             $movie = json_decode(file_get_contents('https://api.themoviedb.org/3/movie/'.$this->id.'?api_key='.config('constants.api_key').'&language=en&append_to_response=recommendations%2Csimilar'), true);
             Recommendation::where(['movie_id' => $this->id])->delete();
